+++ Source/LtInfo.Common/LtInfo.Common.csproj
@@ -178,6 +178,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AssertCustom.cs" />
+    <Compile Include="AthenaPDFUtility.cs" />
     <Compile Include="BootstrapWrappers\BootstrapHtmlHelpers.cs" />
     <Compile Include="CommonUtility.cs" />
     <Compile Include="CommonUtilityTest.cs" />
@@ -202,6 +203,7 @@
     <Compile Include="GeneralUtility.cs" />
     <Compile Include="GeneralUtilityTest.cs" />
     <Compile Include="GeoJson\GeoJsonToSqlGeometryHelper.cs" />
+    <Compile Include="HeadlessChromePDFUtility.cs" />
     <Compile Include="HealthMonitor\HealthCheckStatus.cs" />
     <Compile Include="HtmlHelperExtensions\DropdownForExtensions.cs" />
     <Compile Include="HtmlHelperExtensions\IFieldDefinitionData.cs" />
@@ -248,7 +250,7 @@
     <Compile Include="LtInfoVersionInfo.cs" />
     <Compile Include="Mvc\InMemorySessionStateStore.cs" />
     <Compile Include="Mvc\ISitkaDbContext.cs" />
-    <Compile Include="PDFUtility.cs" />
+    <Compile Include="WkhtmlPDFUtility.cs" />
     <Compile Include="SitkaWebConfiguration.cs" />
     <Compile Include="ModalDialog\ModalDialogFormHelper.cs" />
     <Compile Include="Models\EntityFrameworkHelpers.cs" />
