+++ Source/ProjectFirma.Web/Controllers/ProjectController.cs
@@ -1036,6 +1036,9 @@ Continue with a new {FieldDefinitionEnum.Project.ToType().GetFieldDefinitionLabe
             return RazorPartialView<ConfirmDialogForm, ConfirmDialogFormViewData, ConfirmDialogFormViewModel>(viewData, viewModel);
         }
 
+
+        // Old, using WKHTML to PDF tool.
+        /*
         [ProjectsViewFullListFeature]
         public FileContentResult FactSheetPdf(ProjectPrimaryKey projectPrimaryKey)
         {
@@ -1043,8 +1046,8 @@ Continue with a new {FieldDefinitionEnum.Project.ToType().GetFieldDefinitionLabe
             using (var outputFile = new DisposableTempFile())
             {
                 var pfCookieName = $"{HttpRequestStorage.Tenant.TenantName}_{FirmaWebConfiguration.FirmaEnvironment.FirmaEnvironmentType}";
-                var pdfConversionSettings = new PDFUtility.PdfConversionSettings(CookieHelper.GetAllAuthenticationCookies(Request.Cookies, pfCookieName)) { Zoom = 0.9 };
-                PDFUtility.ConvertURLToPDF(
+                var pdfConversionSettings = new WkhtmlPDFUtility.WkhtmlPdfConversionSettings(CookieHelper.GetAllAuthenticationCookies(Request.Cookies, pfCookieName)) { Zoom = 0.9 };
+                WkhtmlPDFUtility.ConvertURLToPDFWithWkhtml2pdf(
                     new Uri(new SitkaRoute<ProjectController>(c => c.FactSheetForPdf(project)).BuildAbsoluteUrlHttpsFromExpression()),
                     outputFile.FileInfo,
                     pdfConversionSettings);
@@ -1058,7 +1061,30 @@ Continue with a new {FieldDefinitionEnum.Project.ToType().GetFieldDefinitionLabe
                 return File(content, "application/pdf", fileName);
             }
         }
+        */
 
+        [ProjectsViewFullListFeature]
+        public FileContentResult FactSheetPdf(ProjectPrimaryKey projectPrimaryKey)
+        {
+            var project = projectPrimaryKey.EntityObject;
+            using (var outputPdfFile = new DisposableTempFile())
+            {
+                var pfCookieName = $"{HttpRequestStorage.Tenant.TenantName}_{FirmaWebConfiguration.FirmaEnvironment.FirmaEnvironmentType}";
+                // Likely we'll need cookies here too - but let's just get something up first.
+                var pdfConversionSettings = new AthenaPDFUtility.AthenaPdfConversionSettings(CookieHelper.GetAllAuthenticationCookies(Request.Cookies, pfCookieName));
+                Uri factSheetUrl = new Uri(new SitkaRoute<ProjectController>(c => c.FactSheetForPdf(project)).BuildAbsoluteUrlHttpsFromExpression());
+                AthenaPDFUtility.ConvertURLToPDFWithAthena(factSheetUrl, outputPdfFile.FileInfo, pdfConversionSettings);
+
+                var fileContents = FileUtility.FileToString(outputPdfFile.FileInfo);
+                Check.Assert(fileContents.StartsWith("%PDF-"), "Should be a PDF file and have the starting bytes for PDF");
+                //Check.Assert(fileContents.Contains("wkhtmltopdf") || fileContents.Contains("\0w\0k\0h\0t\0m\0l\0t\0o\0p\0d\0f"), "Should be a PDF file produced by wkhtmltopdf.");
+
+                var fileName = $"{project.ProjectName.ToLower().Replace(" ", "-")}-fact-sheet.pdf";
+                var content = System.IO.File.ReadAllBytes(outputPdfFile.FileInfo.FullName);
+                return File(content, "application/pdf", fileName);
+            }
+        }
+        /*
         [ProjectsViewFullListFeature]
         public FileContentResult FactSheetWithCustomAttributesPdf(ProjectPrimaryKey projectPrimaryKey)
         {
@@ -1066,8 +1092,8 @@ Continue with a new {FieldDefinitionEnum.Project.ToType().GetFieldDefinitionLabe
             using (var outputFile = new DisposableTempFile())
             {
                 var pfCookieName = $"{HttpRequestStorage.Tenant.TenantName}_{FirmaWebConfiguration.FirmaEnvironment.FirmaEnvironmentType}";
-                var pdfConversionSettings = new PDFUtility.PdfConversionSettings(CookieHelper.GetAllAuthenticationCookies(Request.Cookies, pfCookieName)) { Zoom = 0.9 };
-                PDFUtility.ConvertURLToPDF(
+                var pdfConversionSettings = new WkhtmlPDFUtility.WkhtmlPdfConversionSettings(CookieHelper.GetAllAuthenticationCookies(Request.Cookies, pfCookieName)) { Zoom = 0.9 };
+                WkhtmlPDFUtility.ConvertURLToPDFWithWkhtml2pdf(
                     new Uri(new SitkaRoute<ProjectController>(c => c.FactSheetWithCustomAttributesForPdf(project)).BuildAbsoluteUrlHttpsFromExpression()),
                     outputFile.FileInfo,
                     pdfConversionSettings);
@@ -1081,5 +1107,29 @@ Continue with a new {FieldDefinitionEnum.Project.ToType().GetFieldDefinitionLabe
                 return File(content, "application/pdf", fileName);
             }
         }
+        */
+
+        [ProjectsViewFullListFeature]
+        public FileContentResult FactSheetWithCustomAttributesPdf(ProjectPrimaryKey projectPrimaryKey)
+        {
+            var project = projectPrimaryKey.EntityObject;
+            using (var outputFile = new DisposableTempFile())
+            {
+                var pfCookieName = $"{HttpRequestStorage.Tenant.TenantName}_{FirmaWebConfiguration.FirmaEnvironment.FirmaEnvironmentType}";
+                var pdfConversionSettings = new HeadlessChromePDFUtility.HeadlessChromePdfConversionSettings(CookieHelper.GetAllAuthenticationCookies(Request.Cookies, pfCookieName));
+                Uri factSheetUrl = new Uri(new SitkaRoute<ProjectController>(c => c.FactSheetWithCustomAttributesForPdf(project)).BuildAbsoluteUrlHttpsFromExpression());
+                HeadlessChromePDFUtility.ConvertURLToPDFWithHeadlessChrome(factSheetUrl, outputFile.FileInfo, pdfConversionSettings);
+
+                var fileContents = FileUtility.FileToString(outputFile.FileInfo);
+                Check.Assert(fileContents.StartsWith("%PDF-"), "Should be a PDF file and have the starting bytes for PDF");
+                //Check.Assert(fileContents.Contains("wkhtmltopdf") || fileContents.Contains("\0w\0k\0h\0t\0m\0l\0t\0o\0p\0d\0f"), "Should be a PDF file produced by wkhtmltopdf.");
+
+                var fileName = $"{project.ProjectName.ToLower().Replace(" ", "-")}-fact-sheet.pdf";
+                var content = System.IO.File.ReadAllBytes(outputFile.FileInfo.FullName);
+                return File(content, "application/pdf", fileName);
+            }
+        }
+
+
     }
 }
